made sep file for chat server
