+++ server/server.js
@@ -9,9 +9,8 @@ const { typeDefs, resolvers } = require('./schemas');
 const db = require('./config/connection');
 const PORT = process.env.PORT || 3001;
 const app = express();
-const socketIo = require('socket.io');
-const { createServer } = require('http');
-const cors = require('cors');
+const CHAT = require('./chatServer');
+// const cors = require('cors');
 // create a new Apollo server and pass in our schema data
 const server = new ApolloServer({
   typeDefs,
@@ -29,7 +28,7 @@ server.applyMiddleware({ app });
 app.use(requestIp.mw())
 app.use(express.urlencoded({ extended: false }));
 app.use(express.json());
-app.use(cors())
+// app.use(cors())
 app.use(function (req, res) {
   const ip = req.clientIp;
   res.end(ip);
@@ -43,27 +42,13 @@ if (process.env.NODE_ENV === 'production') {
 app.get('*', (req, res) => {
   res.sendFile(path.join(__dirname, '../client/build/index.html'));
 });
-const chatApp = express();
-chatApp.use(cors())
-const chatServer = require('http').createServer(chatApp);
-const io = socketIo(chatServer, {
-  cors: {
-    origin: '*',
-    methods: ['GET', 'POST']
-  }
-})
+
 
 db.once('open', () => {
   app.listen(PORT, () => {
     console.log(`API server running on port ${PORT}!`);
     // log where we can go to test our GQL API
     console.log(`Use GraphQL at http://localhost:${PORT}${server.graphqlPath}`);
-  })
-
-});
-chatServer.listen(8080, () => {
-  io.on('connection', (socket) => {
-    socket.emit('message', 'TEST MESSAGE');
-    console.log(socket)
   });
+  CHAT();
 });
